[ios] Check for bool value with char
We convert NSNumbers (and NSStrings) to the appropriate mbgl value
so that we can use NSPredicates to  describe mbgl filters we want
to apply to style layers at runtime.

This change fixes an issue where
the conversion from an NSNumber that represented a bool was not
recognized as such. encode(bool) returns a 'c' or 'b' on 32 bit and
64 bit systems respectively and objCType of an NSNumber representing
a bool always returns 'c'. Now the implementation checks for 'c'
always and NSNumbers representing bool don't fall through and
trigger the exception.