[ios] Use NSDictionary for geojson source options.
This removes a previous implementation of geojson options that used
a new type to transfer the data around. Added in its place is
an options API that takes an NSDictionary that works similarly to
NSAttributedString and many other cocoa APIs that take options.

The implementation of MGLGeoJSONSource checks that values are or the
correct type before attempting to use them